refactor(validator): split validator.ts into orchestrator + cache + step
validator.ts had outgrown its billing as 'the orchestrator' (857 lines,
now 557 after this change). Three responsibility kinds were colocated:

- state-bearing closures (createValidator, compile, getResponseValidator,
  validateRequest, validateResponse, fetch adapters)
- pure per-operation cache construction (cacheFor + helpers)
- pure step functions (validateParameter, validateBody + content helpers)
- the lifted resolveOperationRef primitive

The OperationCache shape already defined the seam between the three.
Split along it:

- packages/validator/src/operation-cache.ts now owns OperationCache /
  ResponseCompiled, buildOperationCache(pathMatch, deps), the shared
  firstContentSchema helper, and resolveOperationRef. The cache builder
  takes its collaborators (resolveRef, compile, compileForDirection) via
  an OperationCacheDeps argument — no closure over createValidator.
- packages/validator/src/validate-step.ts now owns validateParameter
  and validateBody, plus the firstContentMediaType / isJsonMediaType
  helpers. Both step functions are pure given an OperationCache.
- validator.ts keeps createValidator and its closures. cacheFor
  collapses to a WeakMap lookup + buildOperationCache delegate.

No behaviour change: 538/538 tests pass, lint + typecheck green.

Closes #56
